can't delete other posts

+++ backend/app/Http/Controllers/Mypage/PostManageController.php
@@ -73,13 +73,13 @@ class PostManageController extends Controller
     public function destroy(Post $post)
     {
         // 所有チェック
+        if (auth()->user()->isNot($post->user)){
+            abort(403);
+        }
 
         $post->delete();
 
         // 付随するコメントはDBの制約を使って削除する。
-
-
-
         return redirect('mypage/posts');
     }
 }
